+++ src/Filament/Resources/ProductResource/Pages/EditProduct.php
@@ -37,7 +37,7 @@ class EditProduct extends EditRecord
         ];
     }
 
-    protected function getFooterWidgetsColumns(): int|string|array
+    public function getFooterWidgetsColumns(): int|array
     {
         return 1;
     }